+++ hc/lib/emails.py
@@ -41,6 +41,10 @@ def login(to, ctx):
     send("login", to, ctx)
 
 
+def transfer_request(to, ctx):
+    send("transfer-request", to, ctx)
+
+
 def set_password(to, ctx):
     send("set-password", to, ctx)