+++ hc/front/tests/test_copy.py
@@ -16,3 +16,11 @@ class CopyCheckTestCase(BaseTestCase):
         r = self.client.post(self.copy_url, follow=True)
         self.assertContains(r, "This is a brand new check")
         self.assertContains(r, "Foo (copy)")
+
+    def test_it_obeys_limit(self):
+        self.profile.check_limit = 0
+        self.profile.save()
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.copy_url)
+        self.assertEqual(r.status_code, 400)