+++ hc/api/tests/test_status.py
@@ -1,0 +1,12 @@
+from hc.test import BaseTestCase
+
+
+class StatusTestCase(BaseTestCase):
+    url = "/api/v1/status/"
+
+    def test_it_works(self):
+        r = self.client.get(self.url)
+        self.assertEqual(r.status_code, 200)
+
+        self.assertNumQueries(1)
+        self.assertEqual(r.content, b"OK")