+++ hc/settings.py
@@ -28,6 +28,7 @@ def envint(s, default):
 
 
 SECRET_KEY = os.getenv("SECRET_KEY", "---")
+METRICS_KEY = os.getenv("METRICS_KEY")
 DEBUG = envbool("DEBUG", "True")
 ALLOWED_HOSTS = os.getenv("ALLOWED_HOSTS", "*").split(",")
 DEFAULT_FROM_EMAIL = os.getenv("DEFAULT_FROM_EMAIL", "healthchecks@example.org")